Add variable assignment functionality
Added the ability to assign values to variables in the Variable class and updated
the parser and interpreter to support variable assignment. This change includes
updates to the Variable class methods and parsing logic for variable assignments.
Added the option to take input from a file instead of the terminal command line.